+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+
 - <MrkDwn> component for use with <Section> and <Context> ([#73](https://github.com/speee/jsx-slack/issues/73))
 
 ### Changed